requirements.txt
uvicorn
psycopg2-binary
python-dotenv
pandas
sqlalchemy

# Discord
discord.py

# Steam
steammarket
steampy